Restore function order for code review
