+++ core/src/executive/mod.rs
@@ -2,7 +2,6 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
-mod call_create;
 mod context;
 mod executed;
 mod executive;
@@ -30,11 +29,11 @@ impl CollateralCheckResultToVmResult for CollateralCheckResult {
 }
 
 pub use self::{
-    call_create::ExecutiveResult,
     context::InternalRefContext,
     executed::*,
     executive::{
-        contract_address, Executive, ExecutiveGeneric, TransactOptions,
+        contract_address, Executive, ExecutiveGeneric, ExecutiveResult,
+        TransactOptions,
     },
     internal_contract::{
         function, suicide, ABIDecodeError, InternalContractMap,
